+++ packages/featureserver/test/integration/template.json.spec.js
@@ -1,7 +1,8 @@
 const featuresJson = require('../../templates/features.json')
 const fieldJson = require('../../templates/field.json')
 const oidFieldJson = require('../../templates/oid-field.json')
-const { featuresTemplateSchema, fieldsTemplateSchema, oidTemplateSchema } = require('./schemas')
+const layerJson = require('../../templates/layer.json')
+const { featuresTemplateSchema, fieldsTemplateSchema, oidTemplateSchema, layerTemplateSchema } = require('./schemas')
 
 describe('Template content', () => {
   describe('features.json', () => {
@@ -22,6 +23,13 @@ describe('Template content', () => {
     it('should conform to the prescribed schema', () => {
       // Use Joi to build expected schema and test against JSON.
       oidTemplateSchema.validate(oidFieldJson, { presence: 'required' }).should.not.have.property('error')
+    })
+  })
+
+  describe('layer.json', () => {
+    it('should conform to the prescribed schema', () => {
+      // Use Joi to build expected schema and test against JSON.
+      layerTemplateSchema.validate(layerJson, { presence: 'required' }).should.not.have.property('error')
     })
   })
 })